build(scripts): dependency injection for dev.

+++ scripts/generate_queries.py
@@ -40,7 +40,7 @@ def mc_return(env, sim_state, init_action, horizon, policy, max_episodes):
 
         while not done and step_count < horizon:
             with torch.no_grad():
-                obs = torch.tensor(obs).unsqueeze(0)
+                obs = torch.tensor(obs).unsqueeze(0).double()
                 action = policy(obs).data.cpu().numpy()[0]
             obs, reward, done, info = env.step(action.astype("float32"))
             step_count += 1
@@ -66,7 +66,7 @@ def generate_query_states(env, policies, max_transaction_count, args):
                 env_states.append(
                     (obs.tolist(), env.sim.get_state().flatten().tolist())
                 )
-            action = policy(torch.tensor(obs).unsqueeze(0))
+            action = policy(torch.tensor(obs).unsqueeze(0).double())
             noise = torch.normal(0, args.noise, size=action.shape)
             step_action = (action + noise).data.cpu().numpy()[0]
             step_action = step_action.astype("float32")
